+++ include/ReplicationAgent.h
@@ -9,11 +9,12 @@ class ReplicationAgent: public AdvancedAgent {
     private:
         int speed;
         int energy_to_replicate;
+        void find_outbounds_directions(vector <vector <int>> search_positions);
 
     public:
         ReplicationAgent(string name, int max_energy_to_replicate, int x_bound=100, int y_bound=100);
-        void replicate();
         void seek_energy();
+        void replicate();
 };